Adding Interface with status codes
